requirements.txt
ffmpy==0.3.0
python-telegram-bot==13.13
SpeechRecognition==3.8.1